Update Docker build to include Chrome
This commit updates the Dockerfile to pre-download Chrome so that
`pyppeteer` does not each time `aada` is invoked. It uses the
`CHROME_EXECUTABLE_PATH` environment variable to point at the downloaded
binary.

Also a `chrome` user is configured to run the headless browser as it
will not run correctly as `root`.

Lastly a Docker seccomp profile is included for Chrome to grant it
correct permissions on the host. This is lifted wholly from
[jessfraz/dotfiles][0] who has done a lot of work on [Docker seccomp
profiles][1]

[0]: https://github.com/jessfraz/dotfiles/blob/a9803abab0d88cc27a80cac5cd49f20a2419bf2c/etc/docker/seccomp/chrome.json
[1]: https://blog.jessfraz.com/post/how-to-use-new-docker-seccomp-profiles/