Fix use of deprecated _async_send method in pyppeteer
With 3.9 the _async_send method has been deprecated
and causes AADA to print errors in a loop. The fix is
to bump the pyppeteer version used by AADA, and its
dependencies.

Since the API for pyppeteer has changed we also need
to make code changes to AADA. And it seems like we
can just use the pyppeteer Launcher directly without
subclassing it in AADA.

Signed-off-by: itsdalmo <kristian@dalmo.me>

_async_send